my_python/requirements.txt
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
pyyaml>=6.0.0,<7.0.0

